Don't shorten error messages in Eunit Surefire reports
Currently, error messages in Eunit Surefire reports are shortened just
like when written to a terminal.  However, the space limitations that
constrain terminal output do not apply here, so it's more useful to
include the unshortened error message.  Getting the full error
message can be particularly helpful when an assertMatch fails because
of a long and deep error term.